requirements.txt
django
openai
python-dotenv
gunicorn